requirements.txt
appdirs==1.4.0
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
pytesseract==0.1.6
python-Levenshtein==0.12.0
six==1.10.0
